Add broken files
